proj/requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2.4
django-crum==0.7.9
django-dropbox-storage==0.4.1
django-storages==1.11.1
dropbox==11.11.0
idna==2.10
paramiko==2.7.2
Pillow==8.2.0
ply==3.11
psycopg2==2.9.1
pycparser==2.20
PyNaCl==1.4.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
stone==3.2.1
urllib3==1.26.5
